Correct content of collectedAnnounces.
The remove() of a List uses equals to find the element to remove. But in
collectedAnnounce we want to have a correct List of all currently
announced devices, regardless if the devices changed something like IP
addresses or services (UpdateEvent). SO we had to write our own remove()
that uses communicationPath as a criteria if a device is in
collectedAnnounces.